+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Make keyboard shortcuts dialog context aware
 - Add action for cloning a unit with subordinates
 - Simplified standard identity selection.
 - Add zooming and panning to ORBAT charts.